pyproject.toml
[project]
name = "termchat"
version = "0.1.3"
description = "A lightweight CLI ollama frontend."
authors = [
    {name = "SmashedAdam",email = "adam.sy.yeung@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ollama (>=0.5.1,<0.6.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "shellingham (>=1.5.4,<2.0.0)"
]



[tool.poetry.scripts]
termchat = "termchat.main:app"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
